# docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Repository Service for TUF API",
        "description": "Repository Service for TUF Rest API",
        "version": "0.8.0b1"
    },
    "paths": {
        "/api/v1/bootstrap/": {
            "get": {
                "tags": [
                    "v1",
                    "v1"
                ],
                "summary": "Check the bootstrap status.",
                "description": "Check if the bootstrap of the system is done or not.",
                "operationId": "get_api_v1_bootstrap__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BootstrapGetResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "v1",
                    "v1"
                ],
                "summary": "Bootstrap the system with initial signed Metadata.",
                "description": "Submit an asynchronous task to initialize the TUF Repository with initial signed Metadata and Settings. Use the task ID to retrieve the task status in the endpoint /api/v1/task.",
                "operationId": "post_api_v1_bootstrap__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BootstrapPayload"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BootstrapPostResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/config/": {
            "get": {
                "tags": [
                    "v1",
                    "v1"
                ],
                "summary": "List settings.",
                "description": "Returns the configuration settings",
                "operationId": "get_api_v1_config__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/repository_service_tuf_api__config__Response"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "v1",
                    "v1"
                ],
                "summary": "Update settings.",
                "description": "Submit an asynchronous task to update configuration settings. Use the task ID to retrieve the task status in the endpoint /api/v1/task.",
                "operationId": "put_api_v1_config__put",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PutPayload"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PutResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/metadata/": {
            "post": {
                "tags": [
                    "v1",
                    "v1"
                ],
                "summary": "Rotate role metadata.",
                "description": "Submit an asynchronous task to rotate a role metadata that requires offline signing. Use the task ID to retrieve the task status in the endpoint /api/v1/task.",
                "operationId": "post_api_v1_metadata__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MetadataPostPayload"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MetadataPostResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/metadata/sign": {
            "get": {
                "tags": [
                    "v1",
                    "v1"
                ],
                "summary": "Get all metadata roles pending signatures.",
                "description": "Get all metadata roles that need more signatures before they can be used.",
                "operationId": "get_sign_api_v1_metadata_sign_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MetadataSignGetResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "v1",
                    "v1"
                ],
                "summary": "Add a signature for a metadata role.",
                "description": "Submit an asynchronous task to add a signature for a metadata role. Use the task ID to retrieve the task status in the endpoint /api/v1/task.",
                "operationId": "post_sign_api_v1_metadata_sign_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MetadataSignPostPayload"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MetadataPostResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/metadata/sign/delete": {
            "post": {
                "tags": [
                    "v1",
                    "v1"
                ],
                "summary": "Submit a task to delete role metadata in signing process.",
                "description": "Submit an asynchronous task to delete role metadata in signing process. Check the status and result using the task ID and the `get task state` endpoint.",
                "operationId": "post_delete_sign_api_v1_metadata_sign_delete_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MetadataSignDeletePayload"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MetadataSignDeleteResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/artifacts/": {
            "post": {
                "tags": [
                    "v1",
                    "v1"
                ],
                "summary": "Add artifacts to Metadata.",
                "description": "Submit an asynchronous task to add artifacts to Metadata. Use the task ID to retrieve the task status in the endpoint /api/v1/task.",
                "operationId": "post_api_v1_artifacts__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddPayload"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/repository_service_tuf_api__targets__Response"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/artifacts/delete": {
            "post": {
                "tags": [
                    "v1",
                    "v1"
                ],
                "summary": "Submit a task to remove artifacts from Metadata.",
                "description": "Submit an asynchronous task to remove artifacts from Metadata. Use the task ID to retrieve the task status in the endpoint /api/v1/task.",
                "operationId": "post_delete_api_v1_artifacts_delete_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DeletePayload"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/repository_service_tuf_api__targets__Response"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/artifacts/publish/": {
            "post": {
                "tags": [
                    "v1",
                    "v1"
                ],
                "summary": "Submit a task to publish artifacts.",
                "description": "Submit an asynchronous task to publish artifacts not yet published from the RSTUF Database. Use the task ID to retrieve the task status in the endpoint /api/v1/task.",
                "operationId": "post_publish_targets_api_v1_artifacts_publish__post",
                "responses": {
                    "202": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/repository_service_tuf_api__targets__Response"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BootstrapGetResponse": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/GetData"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    }
                },
                "type": "object",
                "required": [
                    "message"
                ],
                "title": "BootstrapGetResponse",
                "example": {
                    "data": {
                        "bootstrap": false
                    },
                    "message": "System available for bootstrap."
                }
            },
            "BootstrapPayload": {
                "properties": {
                    "settings": {
                        "$ref": "#/components/schemas/repository_service_tuf_api__bootstrap__Settings"
                    },
                    "metadata": {
                        "additionalProperties": {
                            "$ref": "#/components/schemas/TUFMetadata"
                        },
                        "type": "object",
                        "title": "Metadata"
                    },
                    "timeout": {
                        "type": "integer",
                        "title": "Timeout",
                        "default": 300
                    }
                },
                "type": "object",
                "required": [
                    "settings",
                    "metadata"
                ],
                "title": "BootstrapPayload",
                "example": {
                    "settings": {
                        "expiration": {
                            "root": 365,
                            "targets": 365,
                            "snapshot": 1,
                            "timestamp": 1,
                            "bins": 1
                        },
                        "services": {
                            "targets_base_url": "http://www.example.com/repository/",
                            "number_of_delegated_bins": 4,
                            "targets_online_key": true
                        }
                    },
                    "metadata": {
                        "root": {
                            "signatures": [
                                {
                                    "keyid": "1cebe343e35f0213f6136758e6c3a8f8e1f9eeb7e47a07d5cb336462ed31dcb7",
                                    "sig": "b3c81b6579442b8672ff460ac36cdb51064109fcfa93f0e72dabbf338f7d0c24e1a91d3696d72d81f60e94b2782ab1bf41e6b8f3e535538216f4379bf30a4302"
                                },
                                {
                                    "keyid": "800dfb5a1982b82b7893e58035e19f414f553fc08cbb1130cfbae302a7b7fee5",
                                    "sig": "fdfbb1877da404504e8b9c111d87680df5291094dafa01f66bfe6031e470d7c63cc477c1e79210a89e27734758e54fb54adc529918b1211d9dd9971c8bb26c0a"
                                }
                            ],
                            "signed": {
                                "_type": "root",
                                "version": 1,
                                "spec_version": "1.0.30",
                                "expires": "2024-02-17T10:55:28Z",
                                "consistent_snapshot": true,
                                "keys": {
                                    "f7a6872f297634219a80141caa2ec9ae8802098b07b67963272603e36cc19fd8": {
                                        "keytype": "ed25519",
                                        "scheme": "ed25519",
                                        "keyval": {
                                            "public": "9fe7ddccb75b977a041424a1fdc142e01be4abab918dc4c611fbfe4a3360a9a8"
                                        }
                                    },
                                    "1cebe343e35f0213f6136758e6c3a8f8e1f9eeb7e47a07d5cb336462ed31dcb7": {
                                        "keytype": "ed25519",
                                        "scheme": "ed25519",
                                        "keyval": {
                                            "public": "ad1709b3cb419b99c5cd7427d6411522e5a93aec6767453e91af921a73d22a3c"
                                        }
                                    },
                                    "800dfb5a1982b82b7893e58035e19f414f553fc08cbb1130cfbae302a7b7fee5": {
                                        "keytype": "ed25519",
                                        "scheme": "ed25519",
                                        "keyval": {
                                            "public": "7098f769f6ab8502b50f3b58686b8a042d5d3bb75d8b3a48a2fcbc15a0223501"
                                        }
                                    }
                                },
                                "roles": {
                                    "snapshot": {
                                        "keyids": [
                                            "f7a6872f297634219a80141caa2ec9ae8802098b07b67963272603e36cc19fd8"
                                        ],
                                        "threshold": 1
                                    },
                                    "timestamp": {
                                        "keyids": [
                                            "f7a6872f297634219a80141caa2ec9ae8802098b07b67963272603e36cc19fd8"
                                        ],
                                        "threshold": 1
                                    },
                                    "targets": {
                                        "keyids": [
                                            "f7a6872f297634219a80141caa2ec9ae8802098b07b67963272603e36cc19fd8"
                                        ],
                                        "threshold": 1
                                    },
                                    "root": {
                                        "keyids": [
                                            "1cebe343e35f0213f6136758e6c3a8f8e1f9eeb7e47a07d5cb336462ed31dcb7",
                                            "800dfb5a1982b82b7893e58035e19f414f553fc08cbb1130cfbae302a7b7fee5"
                                        ],
                                        "threshold": 1
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "BootstrapPostResponse": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/repository_service_tuf_api__bootstrap__PostData"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    }
                },
                "type": "object",
                "required": [
                    "message"
                ],
                "title": "BootstrapPostResponse",
                "example": {
                    "data": {
                        "task_id": "7a634b556f784ae88785d36425f9a218"
                    },
                    "message": "Bootstrap accepted."
                }
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "MetadataPostResponse": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/repository_service_tuf_api__metadata__PostData"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    }
                },
                "type": "object",
                "required": [
                    "message"
                ],
                "title": "MetadataPostResponse",
                "example": {
                    "data": {
                        "task_id": "7a634b556f784ae88785d36425f9a218"
                    },
                    "message": "Metadata Update accepted."
                }
            },
            "MetadataSignDeletePayload": {
                "properties": {
                    "role": {
                        "type": "string",
                        "title": "Role"
                    }
                },
                "type": "object",
                "required": [
                    "role"
                ],
                "title": "MetadataSignDeletePayload",
                "example": {
                    "role": "root"
                }
            },
...
            "PutPayload": {
                "properties": {
                    "settings": {
                        "$ref": "#/components/schemas/repository_service_tuf_api__config__Settings"
                    }
                },
                "type": "object",
                "required": [
                    "settings"
                ],
                "title": "PutPayload",
                "example": {
                    "settings": {
                        "expiration": {
                            "targets": 100,
                            "snapshot": 50,
                            "timestamp": 20,
                            "bins": 5
                        }
                    }
                }
            },
            "PutResponse": {
                "properties": {
                    "data": {
                        "type": "object",
                        "title": "Data"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    }
                },
                "type": "object",
                "required": [
                    "data",
                    "message"
                ],
                "title": "PutResponse",
                "example": {
                    "message": "Settings successfully submitted.",
                    "task_id": "06ee6db3cbab4b26be505352c2f2e2c3"
                }
            },
...
            "repository_service_tuf_api__config__Response": {
                "properties": {
                    "data": {
                        "type": "object",
                        "title": "Data"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    }
                },
                "type": "object",
                "required": [
                    "data",
                    "message"
                ],
                "title": "Response",
                "example": {
                    "data": {
                        "bootstrap": "82281613dba54b8ea88dc86211c77d0a",
                        "root_expiration": 365,
                        "root_threshold": 1,
                        "root_num_keys": 2,
                        "targets_expiration": 365,
                        "targets_threshold": 1,
                        "targets_num_keys": 1,
                        "snapshot_expiration": 1,
                        "snapshot_threshold": 1,
                        "snapshot_num_keys": 1,
                        "timestamp_expiration": 1,
                        "timestamp_threshold": 1,
                        "timestamp_num_keys": 1,
                        "bins_expiration": 1,
                        "bins_threshold": 1,
                        "bins_num_keys": 1,
                        "number_of_delegated_bins": 4,
                        "targets_base_url": "http://www.example.com/repository/",
                        "targets_online_key": true
                    },
                    "message": "Current Settings"
                }
            },
...
        }
    }
}